+++ src/test/java/com/github/jnidzwetzki/bitfinex/v2/test/BitfinexSymbolsTest.java
@@ -6,9 +6,11 @@ import org.junit.Test;
 
 import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexCandleTimeFrame;
 import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexCurrencyPair;
+import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexFundingCurrency;
 import com.github.jnidzwetzki.bitfinex.v2.exception.BitfinexCommandException;
 import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexCandlestickSymbol;
 import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexExecutedTradeSymbol;
+import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexFundingSymbol;
 import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexOrderBookSymbol;
 import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexSymbols;
 import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexTickerSymbol;
@@ -76,5 +78,39 @@ public class BitfinexSymbolsTest {
 	@Test(expected=IllegalArgumentException.class)
 	public void testSymbolCreationInvalid() throws BitfinexCommandException {
         BitfinexSymbols.orderBook(BitfinexCurrencyPair.of("BTC", "USD"), BitfinexOrderBookSymbol.Precision.R0, BitfinexOrderBookSymbol.Frequency.F0, 100);
+	}
+	
+	/**
+	 * Test the funding symbol
+	 */
+	@Test
+	public void testFundingSymbol() {
+		final BitfinexFundingSymbol symbol1 = BitfinexSymbols.funding("USD");
+		final BitfinexFundingSymbol symbol2 = BitfinexSymbols.funding(new BitfinexFundingCurrency("USD"));
+		
+		Assert.assertEquals(symbol1, symbol2);
+	}
+	
+	/**
+	 * Test the funding currency
+	 */
+	@Test
+	public void testFundingCurrency1() {
+		final BitfinexFundingCurrency currency1 = new BitfinexFundingCurrency("USD");
+		final BitfinexFundingCurrency currency2 = BitfinexFundingCurrency.fromSymbolString("fUSD");
+		
+		Assert.assertTrue(currency1.toString().length() > 10);
+		Assert.assertEquals(currency1, currency2);
+		Assert.assertEquals(currency1.hashCode(), currency2.hashCode());
+		
+		Assert.assertEquals("fUSD", currency1.toBitfinexString());
+	}
+	
+	/**
+	 * Test the funding currency
+	 */
+	@Test(expected=IllegalArgumentException.class)
+	public void testFundingCurrency2() {
+		BitfinexFundingCurrency.fromSymbolString("USD");
 	}
 }